+++ StreamProcessor.java
@@ -1,0 +1,20 @@
+import java.io.BufferedReader;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.function.Consumer;
+
+class StreamProcessor implements Runnable {
+    private InputStream inputStream;
+    private Consumer<String> consumer;
+
+    public StreamProcessor(InputStream inputStream, Consumer<String> consumer) {
+        this.inputStream = inputStream;
+        this.consumer = consumer;
+    }
+
+    @Override
+    public void run() {
+        new BufferedReader(new InputStreamReader(inputStream)).lines()
+                .forEach(consumer);
+    }
+}
